# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for archiving files",
        "title": "ZIP Archiver API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/task/new": {
            "get": {
                "description": "Создаёт новую задачу для добавления файловых ссылок и последующего создания ZIP-архива.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать новую задачу архивации",
                "responses": {
                    "200": {
                        "description": "Задача успешно создана",
                        "schema": {
                            "$ref": "#/definitions/new_task.Response"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Превышен лимит одновременно выполняемых задач",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/task/{id}/add": {
            "post": {
                "description": "Добавляет один или несколько файловых URL в существующую задачу архивации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Добавить объекты в задачу архивации",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Список URL-адресов для добавления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/add_objects.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылки успешно добавлены в задачу",
                        "schema": {
                            "$ref": "#/definitions/add_objects.Response"
                        }
                    },
                    "400": {
                        "description": "Задача уже завершена ('Task is completed')",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена ('Task not found')",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Сервис архивации остановлен",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/task/{id}/status": {
            "get": {
                "description": "Возвращает текущий статус задачи архивации, список объектов, ошибки и ссылку на архив (если задача завершена).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить статус задачи архивации",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о задаче",
                        "schema": {
                            "$ref": "#/definitions/get_status.Response"
                        }
                    },
                    "400": {
                        "description": "Параметр taskID отсутствует",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Сервис архивации остановлен",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/zips/{filename}": {
            "get": {
                "description": "Возвращает готовый ZIP-архив задачи по имени файла. Если файл не найден — возвращает ошибку.",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "zips"
                ],
                "summary": "Скачать готовый ZIP-архив",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя ZIP-файла",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ZIP-архив для скачивания",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Файл не найден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "add_objects.Request": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "add_objects.Response": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "get_status.Objects": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                }
            }
        },
        "get_status.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "objects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_status.Objects"
                    }
                },
                "status": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "new_task.Response": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}